Temperature sent to the server

+++ MyHealthVitals/Models/IBluetoothCallBackUpdatable.cs
@@ -7,7 +7,7 @@ namespace MyHealthVitals
 		void SPO2_readingCompleted(int sp02, int bpm, float perfusionIndex);
 		void SYS_DIA_BPM_updated(int bpsys, int bpdia, int bpm);
 		void updatingPressureMeanTime(int pressure);
-		void updateTemperature(int temperature, String type);
+		void updateTemperature(decimal temperature, String type);
 
 		void noticeEndOfReadingSpo2();
 	}
